Jpa repositories

+++ src/main/java/com/hospitalinformationsystem/his/repository/DepartmentRepository.java
@@ -3,5 +3,6 @@ package com.hospitalinformationsystem.his.repository;
 import com.hospitalinformationsystem.his.model.Department;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DepartmentRepository extends JpaRepository<Department, Integer> {
+public interface DepartmentRepository extends JpaRepository<Department, Long> {
+
 }
